+++ eras/conway/impl/src/Cardano/Ledger/Conway/Rules/Utxo.hs
@@ -261,7 +261,8 @@ conwayUtxoTransition = do
     certState
     (utxosGovState utxos)
     tx
-    (updateTreasuryDonation tx utxos)
+    (tx ^. bodyTxL . treasuryDonationTxBodyL)
+    utxos
 
 --------------------------------------------------------------------------------
 -- ConwayUTXO STS